EM-236: Cleanup println, re-organize external pages, account for fixed spelling, minor tweaks.
