+++ functional-tests/src/test/groovy/ComplianceOversightSpec.groovy
@@ -3,7 +3,7 @@ import geb.spock.GebReportingSpec
 import geb.Page
 import pages.app.ComplianceOversightPage
 import pages.app.AuthorizationsPage
-import pages.app.ExternalLinkPage
+import pages.external.ExternalLinkPage
 
 import spock.lang.Unroll
 import spock.lang.Title
