+++ functional-tests/src/test/groovy/LifecycleSpec.groovy
@@ -6,7 +6,7 @@ import pages.app.ReclamationPage
 import pages.app.WaterQualityPage
 import pages.app.TailingsManagementPage
 import pages.app.AuthorizationsPage
-import pages.app.ExternalLinkPage
+import pages.external.ExternalLinkPage
 
 import spock.lang.Unroll
 import spock.lang.Title
@@ -27,6 +27,7 @@ class LifecycleSpec extends GebReportingSpec {
       [tag : "h4", text : "ONGOING LIFECYCLE ACTIVITIES"] | [text : "water quality"]                                                                                                       || WaterQualityPage
       [tag : "h4", text : "ONGOING LIFECYCLE ACTIVITIES"] | [text : "tailings management facilities"]                                                                                      || TailingsManagementPage
       //TODO [tag : "h4", text : "RELATED DOCUMENTS"]            | [text : "Life Cycle of a Mine (Association for Mineral Exploration BC and Mining Association of BC)"]                          || new ExternalLinkPage("", "amebc.ca")
+      //The below link re-directs from "ec.gc.ca" to "canada.ca", which may cause this test to fail if the timing is just right and we assert against the redirect page.
       [tag : "h4", text : "EXTERNAL LINKS & RESOURCES"]   | [text : "Environmental Code of Practice for Metal Mines – Mine Life Cycle Activities (Environment and Climate Change Canada)"] || new ExternalLinkPage("Environmental Code of Practice for metal mines: chapter 2 - Canada.ca", "www.canada.ca")
   }
 
@@ -41,7 +42,7 @@ class LifecycleSpec extends GebReportingSpec {
     where:
       SectionSelector                      | ItemSelector                            || AssertPage
       [tag : "h3", text : "Exploration"]   | [text : "BC Geological Survey"]         || new ExternalLinkPage("Geoscience", "gov.bc.ca")
-      [tag : "h3", text : "Exploration"]   | [text : '“Notice of Work" permit']      || AuthorizationsPage
+      [tag : "h3", text : "Exploration"]   | [text : "“Notice of Work” permit"]    || AuthorizationsPage
 
       //TODO [tag : "h4", text : "Design"]        | [text : "environmental assessment"]     ||
       [tag : "h3", text : "Development"]   | [text : "Mines Act"]                    || AuthorizationsPage
