+++ functional-tests/src/test/groovy/ReclamationSpec.groovy
@@ -3,7 +3,7 @@ import geb.spock.GebReportingSpec
 import geb.Page
 import pages.app.ReclamationPage
 import pages.app.TopicsOfInterestPage
-import pages.app.ExternalLinkPage
+import pages.external.ExternalLinkPage
 
 import spock.lang.Unroll
 import spock.lang.Title
@@ -24,7 +24,7 @@ class ReclamationSpec extends GebReportingSpec {
       //TODO PDF [tag : "h4", text : "RELATED DOCUMENTS"]          | [text : "Mine Reclamation Security in British Columbia Fact Sheet (November 2016)"]                                || new ExternalLinkPage("DRAFT", "gov.bc.ca")
 
       [tag : "h4", text : "EXTERNAL LINKS & RESOURCES"] | [text : "Reclamation & Closure (Ministry of Energy, Mines and Petroleum Resources)"]                               || new ExternalLinkPage("Reclamation & Closure - Province of British Columbia", "gov.bc.ca")
-      [tag : "h4", text : "EXTERNAL LINKS & RESOURCES"] | [text : "Securities (Ministry of Energy, Mines and Petroleum Resources)"]                                          || new ExternalLinkPage("Securities - Province of British Columbia", "eao.gov.bc.ca")
+      [tag : "h4", text : "EXTERNAL LINKS & RESOURCES"] | [text : "Securities (Ministry of Energy, Mines and Petroleum Resources)"]                                          || new ExternalLinkPage("Securities - Province of British Columbia", "gov.bc.ca")
       [tag : "h4", text : "EXTERNAL LINKS & RESOURCES"] | [text : "British Columbia Technical and Research Committee on Reclamation"]                                        || new ExternalLinkPage("February 26, 2018 TRCR | BC Technical and Research Committee on Reclamation", "trcr.bc.ca")
   }
   @Unroll
@@ -37,8 +37,8 @@ class ReclamationSpec extends GebReportingSpec {
       at AssertPage
     where:
       SectionSelector                                                                                            | ItemSelector                                                                            || AssertPage
-      [tag : "h2", text : "Committed to ensuring BC's resources are returned to an environmentally sound state"] | [text : "Health, Safety and Reclamation Code for Mines in British Columbia (the Code)"] || new ExternalLinkPage("Health, Safety and Reclamation Code for Mines in British Columbia - Province of British Columbia", "bc.gov.ca")
-      [tag : "h2", text : "Committed to ensuring BC's resources are returned to an environmentally sound state"] | [text : "Reclamation security"]                                                         || new ExternalLinkPage("Securities - Province of British Columbia", "bc.gov.ca")
+      [tag : "h2", text : "Committed to ensuring BC's resources are returned to an environmentally sound state"] | [text : "Health, Safety and Reclamation Code for Mines in British Columbia (the Code)"] || new ExternalLinkPage("Health, Safety and Reclamation Code for Mines in British Columbia - Province of British Columbia", "gov.bc.ca")
+      [tag : "h2", text : "Committed to ensuring BC's resources are returned to an environmentally sound state"] | [text : "reclamation security"]                                                         || new ExternalLinkPage("Securities - Province of British Columbia", "gov.bc.ca")
 
       //TODO PDF [tag : "h2", text : "Annual Reclamation Reports (ARRs)"]                                                   | [text : "Annual Reclamation Report Requirements"]                                       || new ExternalLinkPage("formatrequirements2016.pdf", "gov.bc.ca")
   }
