--- functional-tests/src/test/groovy/modules/CheckBoxModule.groovy
@@ -1,22 +1,0 @@
-package modules
-
-import geb.Module
-
-class CheckboxModule extends Module {
-    def check() {
-        this.value(true)
-    }
-
-    def uncheck() {
-        this.value(false)
-    }
-
-    def isChecked() {
-        this.value() == true
-    }    
-    
-    def isUnchecked() {
-        !isChecked()
-    }
-    
-}
