+++ functional-tests/src/test/groovy/pages/app/TopicsOfInterestPage.groovy
@@ -10,9 +10,9 @@ class TopicsOfInterestPage extends Page {
     pageTitle { $("#pgTitle").text() }
     HomeBtn { $("#header .brand").click() }
 
-    WaterQualityReadMoreBtn { $("main").$("section").has("h3", text:"Water Quality").$("a").has("span", text:"Read More") }
-    TailingsManagementReadMoreBtn { $("main").$("section").has("h3", text:"Tailings Management").$("a").has("span", text:"Read More") }
-    ReclamationReadMoreBtn { $("main").$("section").has("h3", text:"Reclamation").$("a").has("span", text:"Read More") }
+    WaterQualityReadMoreBtn { $("main").$("section").has("h3", text:"Water Quality").$("a").has("span", text:"Read More").click() }
+    TailingsManagementReadMoreBtn { $("main").$("section").has("h3", text:"Tailings Management").$("a").has("span", text:"Read More").click() }
+    ReclamationReadMoreBtn { $("main").$("section").has("h3", text:"Reclamation").$("a").has("span", text:"Read More").click() }
 
     commonLink { module(CommonLinkModule) }
   }
