+++ functional-tests/src/test/groovy/pages/external/GooglePlusPage.groovy
@@ -1,4 +1,4 @@
-package pages.app
+package pages.external
 
 import geb.Page
